Prevent top level nested attributes
* Add remove_repr macro for testing: This macro removes `#[repr(_)]` from an enum, this is useful for testing
the bug described in #1. `#[remove_repr]` is only available behind the
feature flag test-utils.

* Add a fake `repr` attribute in test-utils: This patch adds a fake `repr` and add tests to try to confuse the
compiler into thinking that it is the real repr. If possible, that would
cause problems with safe-discriminant. That is because we assume that
there is always a repr attributed.

* Reject enums with attribute macros: This patch checks if there is any top level attribute macro expansion,
and if there it, we report it as an error.

* test marcro ordering

fixes: #1

---------

Signed-off-by: Ahmed Abdelraoof <ahmed.abdelraoof@huawei.com>